feat: real time updates with server-sent-events (#15)
Closes #10

Uses Server-Sent events to update frontend for realtime market view.

Because of technical limitations (the mistake to use PHP :smile) it is not very feasible to run SSE directly in the application, because I don't want a PHP process per active user... that's why an external system (Mercure) is used to handle the SSE endpoint. All events are dispatched to Mercure.

TODO:
- [x] make sure updates work smoothly (current state is kinda wonky with content being replaced from actions + events)

Reviewed-on: https://gitea.ascendise.ch/ascendise/php-market/pulls/15
Co-authored-by: Ascendise <ascendise@ascendise.ch>
Co-committed-by: Ascendise <ascendise@ascendise.ch>